medical-report-summarizer/requirements.txt
flask
flask-cors
pytesseract
pillow
pdf2image
openai
python-dotenv
